+++ plugins/tool/git-version/src/main/kotlin/eu/bitfunk/gradle/plugin/tool/gitversion/internal/git/GitLoader.kt
@@ -80,7 +80,7 @@ internal object GitLoader : GitContract.Loader {
             gitDir
         } else {
             scanForRootGitDir(
-                currentRoot.parentFile
+                currentRoot.parentFile,
             )
         }
     }