+++ CHANGELOG.md
@@ -42,7 +42,11 @@ All project changes below.
 
 ## [Unreleased](https://github.com/bitfunk/gradle-plugins/releases/latest)
 
-See [changeset](https://github.com/bitfunk/gradle-plugins/compare/v0.1.0...main)
+See [changeset](https://github.com/bitfunk/gradle-plugins/compare/v0.2.0...main)
+
+## [0.2.0](https://github.com/bitfunk/gradle-plugins/releases/tag/v0.2.0)
+
+See [changeset](https://github.com/bitfunk/gradle-plugins/compare/v0.1.0...v0.2.0)
 
 ### Changed
 
@@ -50,8 +54,11 @@ See [changeset](https://github.com/bitfunk/gradle-plugins/compare/v0.1.0...main)
 
 ### Bumped
 
-- Bitfunk Quality 0.1.0 -> 0.1.2
-- Bitfunk Versioning 0.1.1 -> 0.1.2
+- Bitfunk Plugin Development Convention 0.1.0 -> 0.2.1
+- Bitfunk Quality 0.1.0 -> 0.2.0
+- Bitfunk Versioning 0.1.1 -> 0.2.0
+- Bitfunk Composite Delegator 0.1.0 -> 0.2.0
+- Gradle 7.5.1 -> 8.0
 
 ## [0.1.0](https://github.com/bitfunk/gradle-plugins/releases/tag/v0.1.0)